Increase test expiration TTL
We used to expire the next second, but the next second may
only be a microsecond away, so by the time we actually
re-fetch the value, it could've already been expired.
Let's add another second, now we're certain the value won't
yet be expired by the time we try to fetch it - it will live
between 1 & 2 seconds.